Added Phone Number management

+++ addressbook-api/src/main/java/x/mvmn/learn/java/addressbook/api/model/PhoneNumber.java
@@ -5,5 +5,5 @@ public interface PhoneNumber {
 
 	public int getCountryCode();
 
-	public int getNumber();
+	public long getNumber();
 }
